Left-align text and pad to same length.

Left-align text and pad to same length.

Usage

align_left(
  text,
  width = max(nchar(text)),
  type = c("non-trimmed", "trimmed"),
  sep = " ",
  keep.empty = TRUE
)

Arguments

text

character with strings to be left-aligned.

width

numeric indicating the desired length of strings.

type

character specifying pre-processing. Trims whitespace from text if set to "trimmed".

sep

character specifying the padding. May give unexpected results if not a single character.

keep.empty

logical Indicate if empty strings should be kept empty.

Value

character with left-aligned text.