Sending the scripts to the whirl_r_queue for execution

Sending the scripts to the whirl_r_queue for execution

Usage

internal_run(input, steps, queue, level)

Arguments

input

A character vector of file path(s) to R, R Markdown, Quarto scripts, or files in a folder using regular expression, or to to a whirl config file. The input can also be structured in a list where each element will be executed in parallel.

steps

An optional argument that can be used if only certain steps within a config files is to be executed. Should be equivalent to the names of the steps found in the config file. If kept as NULL (default) then all steps listed in the config file will be executed.

queue

The whirl_r_queue that should execute the scripts

level

Depth of the recursive config calls. The initial call will have 1

Value

A tibble containing the execution results for all the scripts.