whirl Options

Internally used, package-specific options. All options will prioritize R options() values, and fall back to environment variables if undefined. If neither the option nor the environment variable is set, a default value is used.

Checking Option Values

Option values specific to whirl can be accessed by passing the package name to env.

options::opts(env = "whirl")

options::opt(x, default, env = "whirl")

Options

out_formats

Which log format(s) to produce. Possiblities are html, json, and markdown formats:gfm, commonmark, and markua. See also details.
default:

"html"

option:

whirl.out_formats

envvar:

R_WHIRL_OUT_FORMATS (evaluated if possible, raw string otherwise)

track_files

Should files read and written be tracked? Currently only supported on Linux.
default:

FALSE

option:

whirl.track_files

envvar:

R_WHIRL_TRACK_FILES (TRUE if one of 'TRUE', '1', FALSE otherwise)

check_renv

Should the projects renv status be checked?
default:

FALSE

option:

whirl.check_renv

envvar:

R_WHIRL_CHECK_RENV (TRUE if one of 'TRUE', '1', FALSE otherwise)

track_files_discards

List of file naming patterns not be tracked when track_files = TRUE
default:

c("^/lib", "^/etc", "^/lib64", "^/usr", "^/var", "^/opt", "^/sys",
    "^/proc", "^/tmp", "^/null", "^/urandom", "^/.cache", .libPaths())

option:

whirl.track_files_discards

envvar:

R_WHIRL_TRACK_FILES_DISCARDS (as character vector, split on ';' delimiter)

track_files_keep

List of file naming patterns alway to be tracked when track_files = TRUE
default:

paste0("^", getwd())

option:

whirl.track_files_keep

envvar:

R_WHIRL_TRACK_FILES_KEEP (as character vector, split on ';' delimiter)

verbosity_level

How chatty should the log be? Possibilities are quiet, verbose, and debug.
default:

"verbose"

option:

whirl.verbosity_level

envvar:

R_WHIRL_VERBOSITY_LEVEL (evaluated if possible, raw string otherwise)

approved_pkgs_folder

Approved folder library packages
default:

NULL

option:

whirl.approved_pkgs_folder

envvar:

R_WHIRL_APPROVED_PKGS_FOLDER (as character vector, split on ';' delimiter)

approved_pkgs_url

Approved URL library packages
default:

NULL

option:

whirl.approved_pkgs_url

envvar:

R_WHIRL_APPROVED_PKGS_URL (as character vector, split on ';' delimiter)

n_workers

Number of simultanous workers used in the run function
default:

1

option:

whirl.n_workers

envvar:

R_WHIRL_N_WORKERS (evaluated if possible, raw string otherwise)

See also