Internal reuse of options description

Internal reuse of options description

Arguments

verbosity_level

How chatty should the log be? Possibilities are quiet, minimal and verbose. (Defaults to "verbose", overwritable using option 'whirl.verbosity_level' or environment variable 'R_WHIRL_VERBOSITY_LEVEL')

log_dir

The output directory of the log files. Default is the folder of the excuted script. log_dir can be a path as a character or it can be a function that takes the script path as input and returns the log directory. For more information see the examples of run() or vignette('whirl'). (Defaults to dirname, overwritable using option 'whirl.log_dir' or environment variable 'R_WHIRL_LOG_DIR')

n_workers

Number of simultanous workers used in the run function. A maximum of 128 workers is allowed. (Defaults to 1, overwritable using option 'whirl.n_workers' or environment variable 'R_WHIRL_N_WORKERS')

track_files_discards

List of file naming patterns not be tracked when track_files = TRUE (Defaults to c("^/lib", "^/etc", "^/lib64", "^/usr", "^/var", "^/opt", "^/sys", ; "^/proc", "^/tmp", "^/null", "^/urandom", "^/.cache", .libPaths()), overwritable using option 'whirl.track_files_discards' or environment variable 'R_WHIRL_TRACK_FILES_DISCARDS')

track_files

Should files read and written be tracked? Currently only supported on Linux. (Defaults to FALSE, overwritable using option 'whirl.track_files' or environment variable 'R_WHIRL_TRACK_FILES')

out_formats

Which log format(s) to produce. Possiblities are html, json, and markdown formats:gfm, commonmark, and markua. (Defaults to "html", overwritable using option 'whirl.out_formats' or environment variable 'R_WHIRL_OUT_FORMATS')

approved_pkgs_folder

Approved folder library packages (Defaults to NULL, overwritable using option 'whirl.approved_pkgs_folder' or environment variable 'R_WHIRL_APPROVED_PKGS_FOLDER')

check_renv

Should the projects renv status be checked? (Defaults to FALSE, overwritable using option 'whirl.check_renv' or environment variable 'R_WHIRL_CHECK_RENV')

track_files_keep

List of file naming patterns alway to be tracked when track_files = TRUE (Defaults to paste0("^", getwd()), overwritable using option 'whirl.track_files_keep' or environment variable 'R_WHIRL_TRACK_FILES_KEEP')

approved_pkgs_url

Approved URL library packages (Defaults to NULL, overwritable using option 'whirl.approved_pkgs_url' or environment variable 'R_WHIRL_APPROVED_PKGS_URL')