Use whirl

Utility function to setup execution with whirl in your project:

Creates configuration file (default _whirl.yml)

Updates .gitignore to not include log files

See vignette("whirl") for how to specify paths inside the configuration file.

Usage

use_whirl(config_file = "_whirl.yml")

Arguments

config_file

Path to the whirl config file, relative to the project