Whirl R session

Extension of callr::r_session with additional methods for easier creating logs.

Super classes

processx::process -> callr::r_session -> whirl_r_session

Methods

Public methods

whirl_r_session$new()

whirl_r_session$finalize()

whirl_r_session$print()

whirl_r_session$poll()

whirl_r_session$wait()

whirl_r_session$check_status()

whirl_r_session$spin()

whirl_r_session$log_script()

whirl_r_session$create_log()

whirl_r_session$log_finish()

whirl_r_session$create_outputs()

whirl_r_session$clone()

Inherited methods

processx::process$as_ps_handle()

processx::process$format()

processx::process$get_cmdline()

processx::process$get_cpu_times()

processx::process$get_error_connection()

processx::process$get_error_file()

processx::process$get_exe()

processx::process$get_exit_status()

processx::process$get_input_connection()

processx::process$get_input_file()

processx::process$get_memory_info()

processx::process$get_name()

processx::process$get_output_connection()

processx::process$get_output_file()

processx::process$get_pid()

processx::process$get_poll_connection()

processx::process$get_result()

processx::process$get_start_time()

processx::process$get_status()

processx::process$get_username()

processx::process$get_wd()

processx::process$has_error_connection()

processx::process$has_input_connection()

processx::process$has_output_connection()

processx::process$has_poll_connection()

processx::process$interrupt()

processx::process$is_alive()

processx::process$is_incomplete_error()

processx::process$is_incomplete_output()

processx::process$is_supervised()

processx::process$kill()

processx::process$kill_tree()

processx::process$poll_io()

processx::process$read_all_error()

processx::process$read_all_error_lines()

processx::process$read_all_output()

processx::process$read_all_output_lines()

processx::process$read_error()

processx::process$read_error_lines()

processx::process$read_output()

processx::process$read_output_lines()

processx::process$resume()

processx::process$signal()

processx::process$supervise()

processx::process$suspend()

processx::process$write_input()

callr::r_session$attach()

callr::r_session$call()

callr::r_session$close()

callr::r_session$debug()

callr::r_session$get_running_time()

callr::r_session$get_state()

callr::r_session$poll_process()

callr::r_session$read()

callr::r_session$run()

callr::r_session$run_with_output()

callr::r_session$traceback()

Method new()

Initialize the new whirl R session

Usage

whirl_r_session$new(
  verbose = TRUE,
  check_renv = FALSE,
  track_files = FALSE,
  track_files_discards = c(),
  track_files_keep = c(),
  approved_pkgs_folder = c(),
  approved_pkgs_url = c()
)

Arguments

verbose

logical Should the progress be printed to the console?

check_renv

logical Should renv be checked?

track_files

logical Should the files be tracked?

track_files_discards

character Files to discard from tracking

track_files_keep

character Files to keep from tracking

approved_pkgs_folder

character Folder with approved packages

approved_pkgs_url

character URL with approved packages

Returns

A whirl_r_session object

Method finalize()

Finalize the whirl R session

Usage

whirl_r_session$finalize()

Method print()

Print the whirl R session

Usage

whirl_r_session$print()

Returns

invisible self

Method poll()

Poll the process

Usage

whirl_r_session$poll(timeout)

Arguments

timeout

The timeout in milliseconds

Returns

invisible self

Method wait()

Wait for the process to complete

Usage

whirl_r_session$wait(timeout = -1)

Arguments

timeout

The timeout in milliseconds

Returns

invisible self

Method check_status()

Check the status of the callr process

Usage

whirl_r_session$check_status()

Returns

invisible status

Method spin()

Spin the spinner

Usage

whirl_r_session$spin(template = NULL)

Arguments

template

The template to display

Returns

invisible self

Method log_script()

Log the script

Usage

whirl_r_session$log_script(script)

Arguments

script

The script to log

Returns

invisible self

Method create_log()

Create the html log

Usage

whirl_r_session$create_log()

Returns

invisible self

Method log_finish()

Finish the log

Usage

whirl_r_session$log_finish()

Returns

invisible self

Method create_outputs()

Create all log outputs

Usage

whirl_r_session$create_outputs(out_dir, format)

Arguments

out_dir

character Output directory for the log

format

character Output formats to create

Returns

invisible,list of logging information

Method clone()

The objects of this class are cloneable with this method.

Usage

whirl_r_session$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.